src/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.17
setuptools
django-cors-headers==4.6.0
django-filter==24.3
django-polymorphic==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.27.2
drf-writable-nested==0.7.2
Faker==30.8.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.10
# Railway.com deployment dependencies
dj-database-url==2.1.0
whitenoise==6.6.0
Pygments==2.18.0
PyYAML==6.0.2
referencing==0.35.1
rich==13.9.4
rpds-py==0.21.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.7.0
flower==2.0.1
requests==2.31.0
python-dotenv==1.0.0